// src/api/fixtures.ts
export const LM_FIXTURES = {
// ...
  Device: {
    "autoBalancedCollectorGroupId": 1,
    "autoProperties": [
      {
        "name": "addr",
        "value": "127.0.0.1"
      }
    ],
    "autoPropsAssignedOn": 1700000000,
    "autoPropsUpdatedOn": 1700000000,
    "awsState": 0,
    "azureState": 0,
    "collectorDescription": "sample-collectorDescription",
    "containsMultiValue": false,
    "createdOn": 1700000000,
    "currentCollectorId": 1,
    "currentLogCollectorId": 1,
    "customProperties": [
      {
        "name": "addr",
        "value": "127.0.0.1"
      }
    ],
    "deletedTimeInMs": 1700000000,
    "description": "This is a Cisco Router",
    "deviceType": 0,
    "disableAlerting": true,
    "displayName": "Cisco Router",
    "enableNetflow": true,
    "gcpState": 0,
    "hostGroupIds": "16,4,3",
    "hostStatus": "sample-hostStatus",
    "id": 1,
    "inheritedProperties": [
      {
        "name": "addr",
        "value": "127.0.0.1"
      }
    ],
    "isPreferredLogCollectorConfigured": true,
    "lastDataTime": 1700000000,
    "lastRawdataTime": 1700000000,
    "link": "www.ciscorouter.com",
    "logCollectorDescription": "sample-logCollectorDescription",
    "logCollectorGroupId": 1,
    "logCollectorGroupName": "sample-logCollectorGroupName",
    "logCollectorId": 1,
    "name": "Main Collector",
    "netflowCollectorDescription": "sample-netflowCollectorDescription",
    "netflowCollectorGroupId": 1,
    "netflowCollectorGroupName": "sample-netflowCollectorGroupName",
    "netflowCollectorId": 1,
    "op": "sample-op",
    "preferredCollectorGroupId": 1,
    "preferredCollectorGroupName": "sample-preferredCollectorGroupName",
    "preferredCollectorId": 2,
    "relatedDeviceId": -1,
    "resourceIds": [
      {
        "name": "addr",
        "value": "127.0.0.1"
      }
    ],
    "rolePrivileges": [
      "sample-rolePrivileges"
    ],
    "scanConfigId": 0,
    "syntheticsCollectorIds": "1,4",
    "systemProperties": [
      {
        "name": "addr",
        "value": "127.0.0.1"
      }
    ],
    "toDeleteTimeInMs": 1700000000,
    "type": "uptimewebcheck",
    "upTimeInSeconds": 1700000000,
    "updatedOn": 1700000000,
    "userPermission": "sample-userPermission"
  },
// ...
} as const;